kyosint/requirements.txt
pyTelegramBotAPI
django-environ
requests